+++ packages/subscription-controller/src/SubscriptionController.test.ts
@@ -55,7 +55,10 @@ function createCustomSubscriptionMessenger(props?: {
     AllowedEvents['type']
   >({
     name: controllerName,
-    allowedActions: ['AuthenticationController:getBearerToken'],
+    allowedActions: [
+      'AuthenticationController:getBearerToken',
+      'AuthenticationController:performSignOut',
+    ],
     allowedEvents: props?.overrideEvents ?? [
       'AuthenticationController:stateChange',
     ],
@@ -82,9 +85,16 @@ function mockSubscriptionMessenger(overrideMessengers?: {
   const { baseMessenger, messenger } =
     overrideMessengers ?? createCustomSubscriptionMessenger();
 
+  const mockPerformSignOut = jest.fn();
+  baseMessenger.registerActionHandler(
+    'AuthenticationController:performSignOut',
+    mockPerformSignOut,
+  );
+
   return {
     baseMessenger,
     messenger,
+    mockPerformSignOut,
   };
 }
 
@@ -93,10 +103,7 @@ function mockSubscriptionMessenger(overrideMessengers?: {
  *
  * @returns The mock messenger and related mocks.
  */
-function createMockSubscriptionMessenger(): {
-  messenger: SubscriptionControllerMessenger;
-  baseMessenger: Messenger<AllowedActions, AllowedEvents>;
-} {
+function createMockSubscriptionMessenger() {
   return mockSubscriptionMessenger();
 }
 
@@ -129,6 +136,7 @@ type WithControllerCallback<ReturnValue> = (params: {
   initialState: SubscriptionControllerState;
   messenger: SubscriptionControllerMessenger;
   mockService: ReturnType<typeof createMockSubscriptionService>['mockService'];
+  mockPerformSignOut: jest.Mock;
 }) => Promise<ReturnValue> | ReturnValue;
 
 type WithControllerOptions = Partial<SubscriptionControllerOptions>;
@@ -147,7 +155,7 @@ async function withController<ReturnValue>(
   ...args: WithControllerArgs<ReturnValue>
 ) {
   const [{ ...rest }, fn] = args.length === 2 ? args : [{}, args[0]];
-  const { messenger } = createMockSubscriptionMessenger();
+  const { messenger, mockPerformSignOut } = createMockSubscriptionMessenger();
   const { mockService } = createMockSubscriptionService();
 
   const controller = new SubscriptionController({
@@ -161,6 +169,7 @@ async function withController<ReturnValue>(
     initialState: controller.state,
     messenger,
     mockService,
+    mockPerformSignOut,
   });
 }
 
@@ -427,6 +436,16 @@ describe('SubscriptionController', () => {
         expect(mockService.cancelSubscription).toHaveBeenCalledWith({
           subscriptionId: 'sub_123456789',
         });
+      });
+    });
+  });
+
+  describe('triggerAuthTokenRefresh', () => {
+    it('should trigger auth token refresh', async () => {
+      await withController(async ({ controller, mockPerformSignOut }) => {
+        controller.triggerAccessTokenRefresh();
+
+        expect(mockPerformSignOut).toHaveBeenCalledWith();
       });
     });
   });